+++ projects/fhi-angular-components/src/public-api.ts
@@ -14,7 +14,6 @@ export * from './lib/fhi-datepicker/fhi-datepicker.component';
 export * from './lib/fhi-datepicker/fhi-datepicker.module';
 
 export * from './lib/fhi-year-selector/fhi-year-selector.component';
-export * from './lib/fhi-year-selector/fhi-year-selector.module';
 
 export * from './lib/fhi-multiselect/fhi-multiselect.component';
 export * from './lib/fhi-multiselect/fhi-multiselect.module';
@@ -36,4 +35,5 @@ export * from './lib/fhi-week-selector/fhi-week-selector.component';
 export * from './lib/fhi-week-selector/fhi-week-selector.module';
 
 export * from './lib/fhi-time/fhi-weekpicker/fhi-weekpicker.component';
+export * from './lib/fhi-time/fhi-year-range/fhi-year-range.component';
 export * from './lib/fhi-time/fhi-week-range/fhi-week-range.component';