+++ src/MOCK_DB_DATA/library-items/library-items-shared-data.ts
@@ -177,4 +177,8 @@ export const LibraryItemsSharedData: LibraryItemsShared = {
     id: 'time-selector-year',
     title: 'Year selector'
   },
+  TimeSelectorYearRange: {
+    id: 'time-selector-year-range',
+    title: 'Year range'
+  },
 };