+++ src/MOCK_DB_DATA/library-items/time-selectors/_time-selectors.data.ts
@@ -9,6 +9,7 @@ import { TimeSelectorDateRange } from './time-selector-date-range';
 import { TimeSelectorWeek } from './time-selector-week';
 import { TimeSelectorWeekRange } from './time-selector-week-range';
 import { TimeSelectorYear } from './time-selector-year';
+import { TimeSelectorYearRange } from './time-selector-year-range';
 
 const TimeSelectors: LibraryItem[] = [
   ...TimeSelectorDatepicker,
@@ -17,6 +18,7 @@ const TimeSelectors: LibraryItem[] = [
   ...TimeSelectorWeekRange,
   ...TimeSelectorYear,
   ...TimeSelectorDateRange,
+  ...TimeSelectorYearRange
 ];
 
 export const TimeSelectorsData: LibraryItemGroup = {