+++ projects/fhi-angular-components/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 
-> Oct 27, 2023
+> Oct 31, 2023
 
+* :gift: **New** Added Year Range selector
 * :gift: **New** Added Week Range selector
 * :tada: **Enhancement** Change `fhi-btn-flat` -> `fhi-btn-link` to reflect changes in @folkehelseinstituttet/style
 * :bug: **Bugfix** Update deprecated icon class in `fhi-multiselect.component.html`: `icon-xmark` -> `icon-x`